fix unsigned-signed warnings

+++ fhe/test/serialize_test.cpp
@@ -226,7 +226,7 @@ TEST(SERIALIZE_TEST, pubKsKey32) {
     //verify that they are equal
     for (int64_t i = 0; i < int64_t(ks_key->in_params.N); i++) {
         for (int64_t j = 0; j < int64_t(ks_key->l_dec); j++) {
-            for (int k = 0; k < ks_key->out_params.N; ++k) {
+            for (int64_t k = 0; k < int64_t(ks_key->out_params.N); ++k) {
                 ASSERT_LE(log2Diff(ks_key->kskey[i][j].a[0].getAT(k), res->kskey[i][j].a[0].getAT(k)),
                           -int64_t(ks_key->ks_params.fixp_params.torus_limbs * BITS_PER_LIMBS));
